Create About us
Yahweh Spirit of Truth Ministries Church
*About Us*
Welcome to Yahweh Spirit of Truth Ministries Church, a community of believers dedicated to spreading the message of hope,spirit, love, and redemption through YAHWEH. Founded by Pastor Nidanapu Joshua, our ministry is driven by a passion to serve and empower individuals, families, and communities.

Our Mission
- To share the Good News of YAHWEH and His transformative power in people's lives
- To provide a nurturing environment for spiritual growth, discipleship, and fellowship
- To serve the community with compassion, kindness, and generosity

Our Values
- *Faith*: We believe in the power of faith to transform lives and communities
- *Love*: We are committed to demonstrating God's love to a world in need
- *Truth*: We are guided by the truth of God's Word and the leading of the Holy Spirit

Our Vision
- To be a beacon of hope and light in our community, shining the love of Christ to all
- To equip believers with the tools and resources they need to live out their faith
- To partner with other ministries and organizations to advance the Kingdom of God

Meet Our Founder
Pastor Nidanapu Joshua is a dedicated servant of God, called to spread the message of hope, holy spirit and redemption to a world in need. With a heart of compassion and a passion for ministry, Pastor Joshua leads our church with love, wisdom, and integrity.

Join Us
We invite you to join us on this journey of faith and service. Whether you're looking for a community to belong to, or a way to serve others, we welcome you to Yahweh Spirit of Truth Ministries Church. Come and experience the love, joy, and hope that comes from following Jesus Christ.

Contact Us
We'd love to hear from you! Please feel free to contact us with any questions, comments, or prayer requests.

- Email: joshuvanidanapu@gmail.com
- Phone:+91 8328221528
- Address: yahweh spirit of truth miniseries, penugolanu Gampalagudem krisha district ( https://goo.gl/maps/tmrwk1B42ZbKE1T28 )